Added brackets check, `from` constructor, and simplified execution
Added the `from` function for when you want to initialize the `Machine` with a specified length for the `data_tape.`
Implemented a brackets check while parsing the file: now brackets are matched in pairs and an error is returned if some are mismatched.
Simplified adds and subs. Simplified jumping between brackets.
Now `parse_instructions` resets values to support multiple executions.